Moved ASMUtils for consistency between Lama branches
