+++ src/main/java/me/modmuss50/optifabric/patcher/LambadaRebuiler.java
@@ -1,5 +1,6 @@
 package me.modmuss50.optifabric.patcher;
 
+import me.modmuss50.optifabric.util.ASMUtils;
 import net.fabricmc.tinyremapper.IMappingProvider;
 import net.fabricmc.tinyremapper.MemberInstance;
 import org.objectweb.asm.tree.ClassNode;
