+++ src/main/java/me/modmuss50/optifabric/util/ASMUtils.java
@@ -1,4 +1,4 @@
-package me.modmuss50.optifabric.patcher;
+package me.modmuss50.optifabric.util;
 
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.Validate;
